fix: removed unused templates
