+++ apps/website/components/templates/loyalty-program/LoyaltyProgramTemplate.tsx
@@ -1,9 +1,22 @@
-import { Divider, Grid, useMediaQuery, useTheme } from '@mui/material';
+import { useRouter } from 'next/router';
+import { ReactNode } from 'react';
+
+import { TOKENS } from '@gateway/theme';
+
+import ArrowBackIcon from '@mui/icons-material/ArrowBack';
+import {
+  Divider,
+  useMediaQuery,
+  useTheme,
+  Avatar,
+  Box,
+  Grid,
+  IconButton,
+  Stack,
+} from '@mui/material';
 
 import { useWindowSize } from '../../../hooks/use-window-size';
-import MainContentTemplate from './MainContentTemplate';
-import SidebarTemplate from './SidebarTemplate';
-
+import { ClientNav } from '../../organisms/navbar/client-nav';
 type Props = {
   sidebar: React.ReactNode;
   mainContent: React.ReactNode;
@@ -16,6 +29,7 @@ export default function LoyaltyProgramTemplate({
   const theme = useTheme();
   const isMobile = useMediaQuery(theme.breakpoints.down('md'));
   const windowSize = useWindowSize();
+  const router = useRouter();
 
   return (
     <>
@@ -28,9 +42,39 @@ export default function LoyaltyProgramTemplate({
           minHeight: `${windowSize.height}px`,
         }}
       >
-        <SidebarTemplate>{sidebar}</SidebarTemplate>
+        <Grid item xs={12} md={4} sx={{ pt: 2, flexGrow: 0 }}>
+          <Stack
+            direction="row"
+            flexGrow={1}
+            alignItems="center"
+            justifyContent="space-between"
+            gap={1}
+            sx={{
+              display: 'flex',
+              px: TOKENS.CONTAINER_PX,
+              flexGrow: {
+                md: 0.5,
+              },
+              mb: { xs: 2, md: 4 },
+            }}
+          >
+            <IconButton onClick={() => router.back()}>
+              <Avatar>
+                <ArrowBackIcon />
+              </Avatar>
+            </IconButton>
+            <Box sx={{ display: { sm: 'flex', md: 'none' }, mr: 2 }}>
+              <ClientNav />
+            </Box>
+          </Stack>
+          {sidebar}
+        </Grid>
         {!isMobile && <Divider orientation="vertical" flexItem />}
-        <MainContentTemplate>{mainContent}</MainContentTemplate>
+        {!isMobile && (
+          <Grid item xs={12} md sx={{ pt: 2 }}>
+            {mainContent}
+          </Grid>
+        )}
       </Grid>
     </>
   );
