+++ apps/website/components/templates/loyalty-program/components/LoyaltySidebar.tsx
@@ -1,7 +1,6 @@
 import { PartialDeep } from 'type-fest/source/partial-deep';
 
 import { Gates, Loyalty_Program } from '../../../../services/hasura/types';
-import SidebarTemplate from '../SidebarTemplate';
 
 type LoyaltySidebarProps = {
   loyalty: PartialDeep<Loyalty_Program>;
@@ -9,5 +8,5 @@ type LoyaltySidebarProps = {
 };
 
 export function LoyaltySidebar({ gate, loyalty }: LoyaltySidebarProps) {
-  return <SidebarTemplate>Sidebar</SidebarTemplate>;
+  return <>Sidebar</>;
 }
